requirements.txt
asgiref==3.4.1
astroid==2.9.2
backports.zoneinfo==0.2.1
black==21.8b0
click==8.0.1
Django==4.0
django-cors-headers==3.10.1
django-geojson==3.2.0
djangorestframework==3.13.1
environ==1.0
geojson==2.5.0
install==1.3.5
isort==5.10.1
jsonfield==3.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pylint==2.12.2
pytz==2021.3
regex==2021.8.28
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
wrapt==1.13.3